Integrate raft node (#84)
Import the etcd/raft lib module
Write the main loop that operates the RSM;
- Order(block)
- Deliver(block)

Assumptions:
- No dynamic config changes to membership
- Use (as module) the WAL from etcd
- Use (copy) `Storage` from Fabric
- No snapshots: don't create snapshots, don't accept snapshot
- Support node creation
- Support node shutdown and restart
- Test single node only

+++ internal/errors/errors.go
@@ -51,13 +51,13 @@ func (c *ClosedError) Error() string {
 // NotLeaderError is an error that denotes that the current node is not the cluster leader.
 // The error carries the identity of the leader if it is known (>0), or 0 if it is not.
 type NotLeaderError struct {
-	LeaderID int
+	LeaderID uint64
 }
 
 func (n *NotLeaderError) Error() string {
 	return fmt.Sprintf("not a leader, leader is: %d", n.LeaderID)
 }
 
-func (n *NotLeaderError) GetLeaderID() int {
+func (n *NotLeaderError) GetLeaderID() uint64 {
 	return n.LeaderID
 }
